refactor : Coupon과 Member 통합
