+++ src/main/java/com/example/high_five/controller/admin/AdminController.java
@@ -82,4 +82,6 @@ public class AdminController {
 
         return "redirect:/api/coupons/admin/coupons";
     }
+
+    // 여기까지 쿠폰 어드민
 }
