+++ src/main/java/com/example/high_five/dto/point/PointAdminAdjustmentRequest.java
@@ -1,4 +1,4 @@
-package com.example.high_five.dto.coupon;
+package com.example.high_five.dto.point;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PointTransactionResponse {
+public class PointAdminAdjustmentRequest {
     private Long memberId;
-    private Long currentPoint;
+
+    private Long amount;
+
+    private String reason;
 }
