+++ src/main/java/com/example/high_five/dto/point/PointAdminPolicyRequest.java
@@ -1,15 +1,13 @@
-package com.example.high_five.dto.coupon;
+package com.example.high_five.dto.point;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PointAdminPolicyResponse {
+public class PointAdminPolicyRequest {
     private Integer signupPoint;
     private Integer reviewPoint;
     private Integer photoPoint;
