+++ src/main/java/com/example/high_five/dto/point/PointAdminPolicyResponse.java
@@ -1,13 +1,15 @@
-package com.example.high_five.dto.coupon;
+package com.example.high_five.dto.point;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class PointAdminPolicyRequest {
+public class PointAdminPolicyResponse {
     private Integer signupPoint;
     private Integer reviewPoint;
     private Integer photoPoint;
