+++ src/main/java/com/example/high_five/dto/point/PointTransactionResponse.java
@@ -1,4 +1,4 @@
-package com.example.high_five.dto.coupon;
+package com.example.high_five.dto.point;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,10 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class PointAdminAdjustmentRequest {
+public class PointTransactionResponse {
     private Long memberId;
-
-    private Long amount;
-
-    private String reason;
+    private Long currentPoint;
 }
