+++ src/main/resources/templates/admin/coupons.html
@@ -347,6 +347,14 @@
                     </div>
 
                     <div class="form-row">
+                        <label>발급 방식</label>
+                        <select name="couponType">
+                            <option value="NORMAL">일반 (발급함 공개)</option>
+                            <option value="PRIVATE">비공개 (관리자 수동 지급용)</option>
+                        </select>
+                    </div>
+
+                    <div class="form-row">
                         <label>쿠폰 이름</label>
                         <input type="text" name="couponName" placeholder="예: 2025 새해맞이 쿠폰" required>
                     </div>
@@ -389,6 +397,7 @@
                         <th>ID</th>
                         <th>쿠폰 이름</th>
                         <th>설명</th>
+                        <th>발급 방식</th>
                         <th>잔여 / 전체</th>
                         <th>유효 기간</th>
                         <th>상태</th>
@@ -399,6 +408,16 @@
                         <td th:text="${coupon.id}">1</td>
                         <td th:text="${coupon.couponName}">신규 가입 쿠폰</td>
                         <td th:text="${coupon.description}">설명</td>
+                        <td>
+                            <span th:if="${coupon.couponType == 'NORMAL'}"
+                                  style="color: #666;">
+                                일반
+                            </span>
+                            <span th:if="${coupon.couponType == 'PRIVATE'}"
+                                              style="color: #6f42c1; font-weight: bold;">
+                                🔒 비공개
+                            </span>
+                        </td>
                         <td>
                             <span th:if="${coupon.issueCount != null}">
                                 <span th:text="${coupon.remainingCount}" style="color: var(--color-primary); font-weight: bold;">50</span>
